Wizard template: move progress bar to description slot.
The progress bar / step list should be beneath the headline (and beneath status messages, if there are any) but above the form. This is necessary when forms with fieldsets are used: the "fieldset-tabs" should be below the progress bar. The tabs are inserted with javascript as first child in the form.