Only provide things in opengever.testing when [tests] is installed.
This is detected by checking whether the ftw.testing distribution is installed.
The reasion for this is that z3c.autoinclude scans every package in opengever, this includes opengever.testing.
When booting the instance we don't have test extras and therefore it will get import errors (e.g. ftw.testing).