Fixed table display in add-tasktemplate form.
With this changes (https://github.com/4teamwork/ftw.tabbedview/pull/31),
the tabbedview and extjs javascripts are only loaded in tabbedviews. Therefore
we need to include the javscript manually.

This is the easiest but maybe not the best way to solve this issue. But because we will
drop out the tasktemplate feature in the near future it's ok like this. i think.