improve border styling, required to refactor scrolling.
There is still an issue at the top-right corner:
the width has floating point pixels which are rounded differently
per browser and for table-display (header) and block-display (body).

The problem is still visible in chrome with certain window widths but
is reduced to only be visible in some pixels in the corner.