disable QuickInstallerTool events in testing for speed boost.
Patches Products.CMFQuickInstallerTool:
Removes the quickinstaller's subscribers to GenericSetup events
which create snapshots for each installed profile.

The snapshots are used for uninstalling products, which is ususally
not done in tests.

Creating the snapshots consume quite a lot of time.
Disabling it speeds up the testing layer setup time.

We do an early monkey patch so that the event registration still works
and the Plone fixture also profits from the patch.
It is a marmoset patch so that we don't have to cope with function pointers.