Initalize annotations during object creation, to avoid CSRF messages on first object access.
We have to initalize the annotations on every object. To avoid CSRF protection messages on first access of newly created objects, which haven't accessed the annotations during the creation process.

Because the PortletAssignment's, for example `ftw.footer` writes the
PortletAssignments to the objects annotations.

So if the annotations haven't been already accessed, the `AttributeAnnotations` mixin will create the `__annotations__` attribute on the object. This means that not the annotations are marked as changed but rather the object itself. And the we can't whitelist the object itself.