Use clones for headings.
The offset of the sticky headings are calculated once at startup based
on the DOM elements. When the content is getting higher, the position is
no longer valid because the offset is still referring to the
relative positioned elements. So a clone of each heading is being generated. When scrolling,
just the clone is being positioned relative. The offset is calculated
based on the static positioned original heading which has the correct
offset in the DOM.