Don't use grok-style template dir conventions for templates that
aren't supposed to be auto-registered by grok.
The pattern 'viewname_templates/templ.pt' is only needed for templates
that should be auto-associated with a grok.View by grokcore.view's
templatereg.py. If we use this pattern for templates that we use outside
grok auto-registration (e.g. ViewTemplateFile or template="..." in
configure.zcml), grok gets all confused and prints misleading warnings
on startup.