Added temporary patch to add a relation to zc.catalog when adding a object.
The `plone.app.relationfield` package contains a subscriber, which add
relations form behavior fields to the zc.relation catalog. This is needed
because the general z3c.relationfield implementation only works for schema
fields and other field interfaces. But the `plone.app.relationfield` subscriber
only listen on IObjectModifiedEvent, if an object is created with relations
this relations are not registred in the zc.catalog.