Use newly introduce MeetingWrapper as context.
We overwrite the `_getOb` of the Comittee type. This allows us to change the
context for meeting objects to the MeetingWrapper, if they are requested. That
implementation change allow us to register meetings view, as regular
Browserviews without any traversal hack.

So the URL for accessing meetings looks like the following:
 - http://localhost:8080/fd/sitzungen/committee-3/meeting-1/view