CSRF: Add view to enable and disable CSRF tracing monkey patches.
This view allows to enable CSRF tracing, which will monkey patch
the method for object registration on the ZODB connection in
order to save a stack trace at the point a persistent object has
been registered (i.e. changed for the first time in a transaction).

Activating CSRF tracing will patch two methods: register() on the
ZODB connection and _build_csrf_report() on our own
OGProtectTransform. The patch to register() will save a formatted
stack trace to a module global every time a persistent object is
modified for the first time in a transaction. This will then allow
the patched _build_csrf_report() to include that stored stack
trace in the CSRF incident log if a CSRF confirmation dialog is
triggered.

Those patches can also be reverted from this view: The respective
methods will then be restored to their original state.

Additionally, there is a timeout after which the patches will
expire, and remove themselves the next time they're called.