Rework Notitication <-> Watcher relation.
There is no longer a direct relation betweeen Notitication and Watcher, because a Watcher exist for each Actor (could also be a inbox or a group) and a Notitication is generated for exactly one user. Therefore some notitication center features has to be reworked.

Right now the watcher model, has a separte method to resolve the representatives , which differs between users and inbox actors. This is only a temporarilly solution, because the notification improvments should be available in the 4.4 release (backport). This method should use the Actor functionality, but right now it's part of the opengever.core instead of opengever.ogds.models. I'll improve that with a separte pullrequest, which has not be backported to the 4.4 release.