Addded js workaroud to fix notification dropdown menu.
With the PR #1378 the order in the jsregistry gets modified, so that plone's dropdown.js is now called earlier than our notification.js. This means that the notification dropdown menu does not work (don't drop down when clicking on the bell image).

The real problem is that the toggleMenuHandler in plones dropdown.js `return false` instead of just prevent the default link event.

So to resolve this issue, I've implement this workaround.