Reduce conflict errors by using uuid4 in CMFEditions.
Patch for Products.CMFEditions.historyidhandlertool.HistoryIdHandlerTool.register

The default "register" method uses the Products.CMFUid IUniqueIdGenerator
utility for generating the history ID. This utility uses the auto-increment
strategy, which generates a lot of conflicts.

In order to reduce the conflicts when generating the history id,
we switch to the uuid4 implementation, generating a random number instead
and thus not writing to the same place.