no longer use modified_seconds index in navigation
Refactoring of the navigation cache key generation in order to get rid
of the modified_seconds index.

The problem, because of which the modified_seconds was initially
created, is that the default modified index does not sort correctly when
the modification happened in the same minute.

The old approach was to use a more precise modified_seconds index, which
turned out to be not optimal because of conflict errors and updating
issues.

Therefore the approach is switched to the default modified index but in
a way that bad ordering within the same minute is compensated by
considering up to 100 objects but quitting early when it safe in order
to speed up performance.

The new approach has a performance drop in key generation:
- Old implementation: ~0.638ms
- New implementation: ~1.089ms