Refactor sequence number generation.
Refactor sequence number generation in order to reduce conflicts
in content creation requests and to ensure uniqueness.

The old Increaser approach hat various flaws, such as it did not make
sure that the number is unique and it did not really help regarding
conflicts.

The new approach moves the incrementation of the sequence number to a
separate transaction which is commited before the number is returned and
thus uniqueness is assured.
By handling the conflict errors while generating the number we reduce
conflict errors in the main content creation transaction.

The Increaser-objects needs to be migrated to normal integers so that
conflicts arise and we can handle them by retrying.

Up to 10 retries are made before the conflict error is raised to the
main transaction, resulting in a retry of the complete request and
delegating further error handling to Zope so that we have a standard
error handling fallback.