REST API: Customize summary to include translated titles:
This will update the summary's 'title' to be the translated
title (based on the negotiated language, coming from the
request's Accept-Language header) for objects with
ITranslatedTitleSupport.

Note: Because we have combined language codes enabled in the
portal_languages tool, language codes sent in the Accept-Language header
need to be written in the combined form (e.g., 'de-ch') for language
content negotiation to work properly.