Refactord state_filter functionality for tabbedview listings.
The current implementation for the `Active` `All` filter is circuitous and very limited to the current use case. Because we'll extend the dossier listing filters with an additional one, I've had to rework the current implementation.

Therfore I've created new objects Filter and FilterList which should makes the code easier to understand and read.

The changes are only refactoring, the functionality is not changed.