Fix GeverJSONSummarySerializer tests:
With commit bf146b in plone.restapi the behavior for listing folder contents
was changed. It now uses catalog queries instead of obj.objectValues(). This
means, that our repositoryfolders' titles will be fetches from the brains,
not the objects any more.

That in turn means that we'll have the title prefixed with the reference number
in listings, because the pure title isn't available on brains in opengever.
(brain.Title, brain.title_de and brain.title_fr are all prefixed).