Fix: Add GeverTabbedviewDictStorage adapter to handle bumblebee proxy views.
Normally, the tabbedview is storing the last state (i.e. sorting) per user per tab.
Since bumblebee requires a proxy view to determine if the list or the gallery should be
rendered, the mechanism to storing and getting the state was broken. So it was no longer
possible to store and retrive the last table state.

This commit will fix.