Fixed textfiltering for task listings on PostgreSQL.
When using PostgreSQL as OGDS and GlobalIndex Database the tabbedview textfilter has not worked because the sequence_number column is part of the listings searchable_columns. The LIKE operator does not work for Integer columns on PostgresSql, therefore we had to extend the query with a Cast.

Fixes #1884.

Note: Sadly all the tests has no value on working on PostgreSQL, because we use SQLite on the testsetup.