Patch DexterityContent.__getattr__ to correctly fall back to defaults from behavior schemas.
Rationale: The implementation in plone.dexterity 2.1.x grabs *marker
interfaces* from SCHEMA_CACHE.subtypes() for behaviors that have them,
instead of their schema interfaces.

If there's a fallback logic in place (and we can't get rid of it), it
should at least work consistently.

The __getattr__ below is an exact copy of DexterityContent.__getattr__
from plone.dexterity == 2.3.0. That version doesn't use SCHEMA_CACHE at
all for behavior schemata, and so avoids using the questionable
SCHEMA_CACHE.subtypes(). This was fixed in PR 21 in plone.dexterity
as part of a major overhaul / unification of behavior lookups.