Patch DX createContentInContainer to set default values:
createContentInContainer simply ignores default values,
and only persists the exact values that have been passed in. That
issue is hard to notice though, because DX Content.__getattr__ falls
back to returning the field default instead of raising an AttributeError.

So we patch it here to determine the default values for all fields on
a type (including behaviors), and set them for fields that haven't been
provided explicitly.