Use empty list as missing value for relation lists:
There's a bug that's still unfixed:
If a field's value equals the field's missing_value, z3c.form fills
in the field's default, which might be different.

In order to prevent unwanted side effects from this behavior, we
make sure that for relation lists missing value equals the default value.