Have create_restricted_vocabulary() build a VocabularyFactory instead of
a VocabularyFactoryFactory:

A VF is a global utility. Previously these utilities were registered using
utility factories, so create_restricted_vocabulary() was returning a component
factory, that when called, returned a vocabulary factory. By directly registering
the instanciated component instead of a component factory, we can eliminate this
unnecessary factory and help save the planet.