Patch ftw.builder's create() so that it provides IDuringContentCreation
while content creation is in progress.

This is necessary for context aware defaultFactories to correctly
determine whether the context they got passed is the object itself (edit)
or the container (add).

Eventually this will probably make its way into ftw.builder, but at this
point I just need tests to pass to continue with the refactoring, and im
not sure yet whether this will be the final mechanism I'll be going with.