Implement UserTurboIndex.
When sorting a tabbed view by a user column, we had to sort the complete
result set using the brain metadata in order to have it sorted by
fullname rather than userid.
The problem is that this breaks lazyness and triggers brain builds for
all items in the result set, leading to performance problems.

In order to fix that, we are adding a new UserTurboIndex, currently
replacing the "Creator" and "responsible" indexes.

This new index allows to sort by another value than the indexed value,
by providing a "documentToKeyMap".
The catalog does only use that though when the "limit" in the sort
method is not None, we therefore have to patch this argument.