Fix rendering of new agenda points in IE11.
Closes https://github.com/4teamwork/opengever.core/issues/2437

IE11 does not render newly added agenda points in a new tab before an explicit reload.
This behavior is documented under http://go.microsoft.com/fwlink/?LinkID=291337.

So explicitly refresh the page after the `pageshow` event.
The pageshow event is fired when a session history entry is being traversed to.
This includes back/forward as well as initial page-showing after the onload event.

For some reasons the `persisted` attribute on the event is not set properly:
https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/101518/
So the agenda points are re-rendered on every pageload.