testing: install opengever.policy.base:default.
Change the test setup so that we install opengever.policy.base:default
rather than its dependencies, so that we include the configuration from
the policy profile.

The language is reset to English so that we do not have to change all
the tests.

Because a bunch of tests rely on having ``filter_content_types=False``
configured for the ``Plone Site`` FTI, we still keep that for now.
This should be awoided for future tests and can be elminiated step by
step.