Fix keyword widget installation order.
opengever.base:default does reconfigure the rolemap configuration of
ftw.keywordwidget.
But because ftw.keywordwidget was installed _after_ opengever.base, the
customization was overwritten with the defaults on a fresh installation.

It did work in test, because the tests do not install GEVER exactly as
it is installed in production.
It did work when upgrading, because the upgrade step did it in the right
order.

The problem by verified by checking the security configuration on the
Plone site (manage_access): it is correctly configured when the
Contributor-role has the "ftw.keywordwidget: Add new term" permission on
the site root.