Include Hotfixes in development buildout.
In order to get the ${buildout:hotfix-eggs} variable populated
with the appropriate hotfixes, the extends order of
plone-development.cfg and test-plone-4.3.x.cfg had to be switched,
because plone-development.cfg resets/initializes that variable.

This cause ${buildout:always-checkout} = force to be inherited from
the testing cfg into our GEVER development buildout, which is why
we need to switch that back to 'false' here.