Allow certain z3c forms to be prefilled via GET request:
This is necessary because with the introduction of
Products.PloneHotfix20160830, forms (which are POSTed)
are not allowed to be prefilled via GET requests any more
(by default).

By setting allow_prefill_from_GET_request = True on forms that
need this behavior, prefilling via GET can be whitelisted on a
per-form basis.