Meeting: add a "Word implementation" feature flag.
The new feature flag "Word implementation" is used for switching the
meeting implemetation to the new Word implementation.

The new word implementation allows to write protocols directly in word.
In order to develop it in parallel it needs to be feature flag.

When enabling the Word implementation feature flag no meeting objects
should exist, otherwise they migth lack a proper migration at this point.