get rid of integration testing.
Reasons for removing integration testing:
- The integration testing layer does not isolate the database on the
test level and should therefore be awoided.
- Having functional testing as well as integration testing only adds
setup time but does not give a benefit at all.
- The one and only existing integration test is skipped because it does
not work anymore. It is a mock test case, which is bad in its own way,
and it wont work. Keeping it around has no benefit for us.