Disable editing of documents in proposal which have been submitted.
When a proposal is submitted, the original porposal must not be edited
anymore.
Since the proposal body text is now stored in a document within the
proposal, we need to introduce a new DC workflow state so that we can
manage the permissions for the file and lock the file.