Do not sync task workflow changes to forwarding predecessors.
Workflow changes should never be synced to a forwarding predecessor, because a forwarding predecessor is always closed and stored in a yearfolder. The forwarding has been already ignored by syncing, but the special handling has been removed by mistake during the rework PR #2828.

This fixes an `Forbidden` exception when trying to reassign a task, which has been created while assigning a forwarding to a dossier.