Add integration testing layer with fixture.
The integration testing layer sets up a fixture with commonly used
objects.
By setting up a fixture instead of creating all the objects in each test
we want to speed up the test suite.

Since we cannot rollback commits in SQL, we use integration testing and
avoid unwanted commits so that we can achieve isolation between tests by
rolling back to savepoints.

This change provides the infrastructure for building the fixture and
gradually switching tests to the new integration testing.

It is a conscious decision to call the new testing "integration testing"
and the old testing "functional testing" in order to have a simple
language to communicate between developers.