Fix mail title initalization
In the newer plone version (4.3.14) the behavior of creating an object has changed, therefore the mail titled gets prefilled with the default_title fallback (see https://github.com/4teamwork/ftw.mail/blob/master/ftw/mail/mail.py#L45). So we need to notice this fallback, when checking if a title exists or not in the `initalize_title` event handler.