Support plone.protect's auto CSRF protection as good as possible.
The problem is that we use the same connection and transaction for
preparation as for performing a request with ftw.testbrowser.

This means that we may already have changed objects on the connection
but the change is not from within the request.

We fix that by marking all objects which are already marked as changed
on the current as safe for CSRF.
This also means that the auto protection does no longer trigger within
the test for the followed requests.