Move sphinx versions into versions.cfg.
Make sure that all versions are pinned and that all checks and builds
share the same versions.cfg.
Having all versions pinned is important in order to be able to reproduce
an older state.
Having all versions in the same file is good for avoiding conflicting
versions when combining multiple buildout configs.