Let long tests fail.
We need to make sure that we optimize tests so that they are fast.
Most tests shouldn't take that long when implemented correctly.
Therefore we want to fail whenever a test takes longer than the
threshold.

Exceptions can be made with a decorator where the new time limit and a
reason must be provided.