Fix unicode error in committee group vocabulary.
The token of a vocabulary term must be a bytestring because it will be
sent over HTTP. The VocabularyTerm class casts tokens to bytestring in
it's __init__, which will throw an encoding error when the token is a
unicode containing non-ASCII characters.

Simply encoding the unicode as utf-8-bytestring will not work because
the z3cform validation will fail and not allow to select the term.

A good token does only contain ASCII characters while beeing
reproducable and hopefully not conflicting with other terms.
Therefore I changed the token to be a normalized string generated from
the group ID.

It seems that groups usually have no title. We should not fallback to
the normalized string then but to the original group id.