Fix repository folder in examplecontent committees.
The configured repository folder in a committee is used for storing
meeting content in dossiers.
Since the meeting module creates dossiers in this folder, the configured
folder must be a leaf-node.

This change updates the examplecontent from using branch-nodes to
leaf-nodes, so that it is possible to create a meeting in a those
committees.