Enable *.msg download for all download copy links.
Also in the tooltip and bumblebeeoverlay. The `get_download_copy_tag` was unnecessary overwritten in both overviews (Document and Mail) but did exactly the same. I dropped those methods in the overviews and moved the special handling for the original_message to the actionbuttons-mixin itself. So the `copy download` action works everywhere in the same manner.