Handle unknown asynchronous tooltip response.
The tooltip may call the content with a logged out user. This leads
to an unwanted rendering of the whole login page in the tooltip.

So add a response header to make sure the response is a true tooltip response
and redirect the user to the login page with a simple reload.