Cache integration testing setup
When developing ``opengever.core``, a developer often runs single test module,
with ``bin/test -m opengever.dossier.tests.test_activate`` for instance.
This will set up a complete fixture each time.
In order to speed up the feedback loop when developing,
we try to cache the database after setting up the fixture.
This will speed up the test runs, but it also makes the result inaccurate:
if the cachekeys do not detect a relevant change, we may not realize
that something breaks.

Because the results are not accurate and this is an experiment, the feature is
considered experimental and therefore disabled by default.

You can enable the feature by setting an environment variable:

    GEVER_CACHE_TEST_DB=true bin/test -m  opengever.dossier.tests.test_activate

You can manually remove / rebuild the caches:

    ./bin/remove-test-cache

This feature is disabled on the CI server.