Register Administrator- and Publisher-roles with lawgiver.
Use the lawgiver:role directive for registering Administrator- and
Publisher-roles.

The old sharing page role utilites protected the roles with the
standard "Sharing page: delegate roles" permission.
Each role should have its own permission; the standard permission is
used for protecting the sharing view.

In order to have the same behavior as before, all roles which used to
have the "Sharing page: delegate roles" permission and therefore were
able to delegate Administrator and Publisher are receiving the
role-specific permissions on the site root (rolemap.xml).
Since the standard permission is required too, the current workflows
still have the same effect.

The document workflow also grants the standard delegate permission to
the Owner role, but we discussed that this is an error.