Sharing: let lawgiver configure the visible roles.
Lawgiver configures the "visible roles" by configuring the
role-specific delegate permissions properly.

This means we do not need to customize the list of visible roles in
the view as it is configured through security.

When a context has a lawgiver workflow, we fallback to displaying all
roles where the user has the delegate permission.
This will also list all newly added roles without the need of changing
the sharing view.