Remove sharing unit / mock test.
The test mainly tests the configuration of visible roles in the
sharing view. We should not test configuration.
It is also mock test, so it does most likely not test the real world
but some mocks.

I've removed it because a lawgiver specification check was inserted
in the sharing view implementation. This requires access to
portal_workflow. This would end up in another mocking mess.