Fixture: improve documentation and assert readme updated.
1) Fix the three in the fixture documentation, so that it works with real
Plone objects as well as SQL wrappers.
The problem was that some were prefixed with the Plone site path,
others were relative to the Plone site.
Since the IntegrationTestCase traverses on the Plone site it is
correct to make all paths relative to the Plone site.

2) Let the test fail if the readme was udpated when running the test.
This does only fail once and then leave an uncommitted, changed readme
behind, which must be committed be the developer.
On the CI-server it will fail when a PR changes the fixture but misses
the readme update.