Disable ExtJS in integration testing.
ExtJS is implemented in JavaScript, but ftw.testbrowser does not
support JavaScript.

In order to test listing tabs anyway, ExtJS is deactivated for
integration tests so that ftw.tabbedview renders a simple table,
which is testable with ftw.testbrowser.

ExtJS is also deactivated in functional testing, so this will behave
the same.