Word meeting: update styling and placement of meeting view components.
The metadata is moved from the sidebar to a top info section (3
columns), so that it takes less space, especially when scrolling down
to the agenda items.
For clearer communication, icons are replaced with text labels.

The left sidebar will be removed when everything is moved to a place.
At the left hand side we will have a navigation instead, for
navigating between the agenda items.

The "add" section is moved to the bottom (because that's where the
items will appear), and the labels are improved.