Also assign Editor when assigning CommitteeResponsible.
CommitteeResponsible is assigned automatically, don't expose it in the
@@sharing tab but only mention that this role is assigned by the
system.

This was removed in a previous commit, but is necessary at the moment.
The document workflow currently is implemented in such a way that it interrupts
inheritance for permissions vital to checkout/checkin. THe only feasible way
(at the moment) is to also assign users the Editor role to allow document
editing. If we want to change this behavior the document workflow must be changed
which is not recommended at the moment.