Fix shared default value for excerpt lists.
When creating a new excerpt for a submitted proposal and setting the
excerpt list attribute/field on the submitted proposal the default
value from the ISubmittedProposal schema was used instead of a new list
instance. This happened due to usage of getattr which already returns
the schema default. This lead to that that default value being shared
between multiple instances of submitted proposals.

We attempt to fix this with an upgrade step. Excerpts are stored in the
meeting-dossier. We fix shared excerpt lists by creating a new list instance
and filtering out the excerpts that come from foreign meeting dossiers
should they be present in the excerpt list.

We can't completely fix relations within the same meeting, so it may happen
that multiple submitted proposals point to the same excerpt within the same
meeting dossier. Since this issue never entered production that should be
rare and unproblematic though.