OGGBundle: Fix title encoding in constructor section.
Currently the construtor section does not ensure the encoding for
the title attribute, which must be unicode. This can lead in to broken
`zope.i18nmessageid.message.Message` objects in the journal of
an object, which raises an `UnicodeDecodeError` (see
https://sentry.4teamwork.ch/sentry/onegov-gever/issues/3941/) when
trying to translate the Message by `zope.i18n.translate`.

Therefore the journal of those object can not be rendered or converted
in to a PDF when resolving the Dossier.