Register an is_subdossier indexer for dossiertemplates.
We querying dossiertemplates with an `is_subdossier=False` query in
the dossiertemplate listing, but there is no indexer for
dossiertemplates.

With a former change which has changed the `is_subdossier` index from a FieldIndex
to a BooleanIndex, this leads to a confusing behavior when switching for the
dossiertemplate listing tab to the template document tab and
back.