Fix ordering in bumblebee gallery.
Content listings should be ordered in a way the user can interpret.
Currently we order by the catalog's internal values, such as the rid,
which is confusing for users and considered a bug.

It's best to order by modified date by now, in order to show the
latest changes, so that users can find recently used documents easily.

In the future we will hopefully implement an ordering menu, where the
user can decide how to order. But modified date is a good fix for now.