Tabbedview: let the catalog sort date indexes.
ftw.table does not sort by date indexes in order to circumvent the
problem that date indexes with falsy values remove result from the
result set.

ftw.table instead sorts the results afterwards, breaking the lazyness
and retrieving the brains. This is approach is years old and not a
good solution.

The better solution to avoid such problems is to make sure that a date
index never contains falsy values. For optional dates this could easily
achieved by using an early (1900) or late (2199) default date in the
indexer.
The catalog index will decide whether to use an early or late default
date because it does not have this competence.