SPV: Fix dialog text when closing a meeting.
When closing a meeting, the dialog used to claim that the protocol
cannot be edited after closing the meeting.

This is actually wrong:
The protocol is locked and can not be edited when the meeting is
_open_.
Closing the meeting in fact _unlocks_ the document and makes it
possible to edit it.