Fix caching problems in JSON responses with IE 11.
The JSONResponse component is used for generating standardized JSON
responses of read-only (GET) as well as writing (POST) requests.

The responses were not equipped with caching headers (such as
cache-control). If a response does not have a cache-control header,
some browsers, such as IE 11, may cache the response, at least of GET
requests.

This component is primarely used for AJAX requests. We usually want to
avoid caching AJAX responses as their content may change over time.

We especially do not want to let the browser be in control of cache
invalidation.

Therefore we use no-cache headers on all AJAX responses by default.