Allow to set meeting excerpt titles
A user can generate several excerpts, to send them to different
institutions. Automatically generated excerpts share a name, making it
very difficult for the user to determine which excerpt is meant for
which institution. This problem is solved by allowing the user to set an
excerpt title for each created excerpt.