Trash document of deleted ad hoc agenda items
When a responsible of a committee creates an agenda item ad hoc, a
document is created to contain the ad hoc item's proposal / decision.
If the ad hoc item is removed, the related document should be trashed.

If the agenda item is created from a proposal (handed in by a proposer),
then the proposal should remain until a related agenda item has been
scheduled.
Therefore removing the related agenda item should keep the document
untrashed.