Fix private folder creation for userids containing dashes.
The current PrivateFolder implementation assumes that the folder id is
exactly the same than the id of the corresponding user. But the
portal_membership tool, convert the id (for example `-` gets
replaced with `--`, see
https://github.com/plone/Products.PlonePAS/blob/master/src/Products/PlonePAS/tools/membership.py#L318)
therefore this assumption is wrong and leads to the following error
when trying to create a memberarea for userid containing dashes:

```
  File "/plone.api-1.7-py2.7.egg/plone/api/user.py", line 373, in grant_roles
    raise InvalidParameterError('User could not be found')
InvalidParameterError: User could not be found
```

So I've changed the implementation and using `getOwner()` to get the
corresponding userid of an private folder.