Remove excerpt template from forms
The excerpt template is not available when the word-meetin feature
enabled, and is therefore hidden.
Since the field is required (when the word-meeting feature is not
activated) the require-argument is changed by a constraint.
The constraint checks if there's a given value, which is (almost)
equivalent to setting required=True.