./bin/i18n-build: require --all for building all packages.
Developers usually work on a specific package and want to rebuild this
package, not all translations.
Running bin/i18n-build without arguments used to build all packages
and was hard to abort because of the subprocesses, leaving the
developer with many unwanted changes when executing accidentally.
Therefore it is now required to explicitly use --all or a specific
package name.