Fix "stale association proxy" error.
Fixes a flaky bug where the use of get_watchers sometimes threw a
"InvalidRequestError: stale association proxy, parent object has gone out of scope"

The reason for this error is that resource.watchers is an association
proxy, which is implemented lazy.
This leads to situations where the garbage collector removes objects
which are parts of lambdas within the association proxy object,
because the GC believes there are no references to those
objects (refcount is 0).

This means we need to consume the watchers within the same scope we
access the watchers attribute. Thus the get_watchers method cannot
return a lazy association proxy object but must resolve it first.

This is achieved by putting the watchers into a tuple. This resolves
lazyness while making sure that no user of get_watchers tries to add
new watchers (expecting an association proxy).